refactor: move code used by one or more Table classes to a mixin
