--- src/3rdparty/walkontable/src/overlay/table/master.js
@@ -1,7 +1,0 @@
-import Table from '../../table';
-
-class MasterTable extends Table {
-
-}
-
-export default MasterTable;
