+++ src/3rdparty/walkontable/src/table/_base.js
@@ -1,4 +1,4 @@
-import Table from '../../table';
+import Table from '../table';
 
 class OverlayTable extends Table {
   alignOverlaysWithTrimmingContainer() {
