+++ src/3rdparty/walkontable/src/table/mixin/stickyRowsBottom.js
@@ -1,69 +1,50 @@
-import OverlayTable from './_base';
+import { defineGetter } from '../../../../../helpers/object';
 
-class BottomLeftCornerOverlayTable extends OverlayTable {
+const MIXIN_NAME = 'stickyRowsBottom';
+
+/**
+ * @type {Object}
+ */
+const stickyRowsBottom = {
   getFirstRenderedRow() {
     const index = this.wot.getSetting('totalRows') - this.wot.getSetting('fixedRowsBottom');
     if (index < 0) {
       return -1;
     }
     return index;
-  }
+  },
 
   getFirstVisibleRow() {
     return this.getFirstRenderedRow();
-  }
-
-  getFirstRenderedColumn() {
-    const totalColumns = this.wot.getSetting('totalColumns');
-    if (totalColumns === 0) {
-      return -1;
-    }
-    return 0;
-  }
-
-  getFirstVisibleColumn() {
-    return this.getFirstRenderedColumn();
-  }
+  },
 
   /**
    * @returns {Number} Returns -1 if no row is visible, otherwise source index of the last rendered row
    */
   getLastRenderedRow() {
     return this.wot.getSetting('totalRows') - 1;
-  }
+  },
 
   /**
    * @returns {Number} Returns source index of last visible row
    */
   getLastVisibleRow() {
     return this.getLastRenderedRow();
-  }
-
-  getLastRenderedColumn() {
-    return this.getRenderedColumnsCount() - 1;
-  }
-
-  getLastVisibleColumn() {
-    return this.getLastRenderedColumn();
-  }
-
-  getRenderedColumnsCount() {
-    const totalColumns = this.wot.getSetting('totalColumns');
-    return Math.min(this.wot.getSetting('fixedColumnsLeft'), totalColumns);
-  }
+  },
 
   getRenderedRowsCount() {
     const totalRows = this.wot.getSetting('totalRows');
     return Math.min(this.wot.getSetting('fixedRowsBottom'), totalRows);
-  }
+  },
 
   getVisibleRowsCount() {
     return this.getRenderedRowsCount();
   }
+};
 
-  getVisibleColumnsCount() {
-    return this.getRenderedColumnsCount();
-  }
-}
+defineGetter(stickyRowsBottom, 'MIXIN_NAME', MIXIN_NAME, {
+  writable: false,
+  enumerable: false,
+});
 
-export default BottomLeftCornerOverlayTable;
+export default stickyRowsBottom;
