+++ src/3rdparty/walkontable/src/table/mixin/stickyRowsTop.js
@@ -1,40 +1,50 @@
-import OverlayTable from './_base';
+import { defineGetter } from '../../../../../helpers/object';
 
-class BottomOverlayTable extends OverlayTable {
+const MIXIN_NAME = 'stickyRowsTop';
+
+/**
+ * @type {Object}
+ */
+const stickyRowsTop = {
   getFirstRenderedRow() {
-    const index = this.wot.getSetting('totalRows') - this.wot.getSetting('fixedRowsBottom');
-    if (index < 0) {
+    const totalRows = this.wot.getSetting('totalRows');
+    if (totalRows === 0) {
       return -1;
     }
-    return index;
-  }
+    return 0;
+  },
 
   getFirstVisibleRow() {
     return this.getFirstRenderedRow();
-  }
+  },
 
   /**
    * @returns {Number} Returns -1 if no row is visible, otherwise source index of the last rendered row
    */
   getLastRenderedRow() {
-    return this.wot.getSetting('totalRows') - 1;
-  }
+    return this.getRenderedRowsCount() - 1;
+  },
 
   /**
    * @returns {Number} Returns source index of last visible row
    */
   getLastVisibleRow() {
     return this.getLastRenderedRow();
-  }
+  },
 
   getRenderedRowsCount() {
     const totalRows = this.wot.getSetting('totalRows');
-    return Math.min(this.wot.getSetting('fixedRowsBottom'), totalRows);
-  }
+    return Math.min(this.wot.getSetting('fixedRowsTop'), totalRows);
+  },
 
   getVisibleRowsCount() {
     return this.getRenderedRowsCount();
   }
-}
+};
 
-export default BottomOverlayTable;
+defineGetter(stickyRowsTop, 'MIXIN_NAME', MIXIN_NAME, {
+  writable: false,
+  enumerable: false,
+});
+
+export default stickyRowsTop;
